Add a test for the type unification code.
This compares the list of types present in the parsed LLVM module
against those reported by a plain dump using 'opt'.  Types that
are unified only appear once and without the .NNN suffix added by
the LLVM linker.

This isn't particularly useful by itself, but is a good regression
tester.